Make Walk.Builder mutable again.
Turns out there is little reason for immutability here, because the act of
traversing can't be easily parallelized, at least with the traversal
implementations that will be done here. Those implementations will be using
internal stacks of Iterators during traversal, and by definition, traversal
in BFS/DFS is in a particular order.

However, that doesn't mean the products of the traversal (the Walks) can't
be processed in parallel. Those are immutable, and so that is safe.