Refactor the traversal and walk builder innards.
Make Walk.Builder immutable (again). This is so that it
can be used in a breadth-first traverser, which does not
work with a stack.

This pushed the depth-first traversers (all 3) to change to
keep track of a stack of (iterator/walk-builder) pairs
instead of just a stack of iterators.

It turns out this actually results in cleaner code as well,
fewer objects, and more immutability. So this was
definitely the right direction.