Refactor Traversals.preOrder() to be consistent with Gauva TreeTraverser.
This is the beginning of a major refactoring to include all the traversals.
The goal is to decouple the tracking of Walks from the traversal. All
traversals will use a basic TreeTraverser<T>, except for the leaf/element
traversals. If tracking Walks is desired, then the client should provide
a TreeTraverser<Walk> as the adjacency function, to either these traversals
or the ones in Guava.